LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-events.optional-expected.txt


one
two
three
 Click

FAIL Events, implicit button assert_array_equals: after showing, events from test_driver.click expected property 4 to be "focusin on option1 at select. open: true." but got "focusin on select at select. open: false." (expected array ["pointerdown on select at select. open: false.", "pointerdown on select at wrapper. open: false.", "mousedown on select at select. open: false.", "mousedown on select at wrapper. open: false.", "focusin on option1 at select. open: true.", "focusin on option1 at wrapper. open: true.", "pointerup on select at select. open: true.", "pointerup on select at wrapper. open: true.", "mouseup on select at select. open: true.", "mouseup on select at wrapper. open: true.", "click on select at select. open: true.", "click on select at wrapper. open: true."] got ["pointerdown on select at select. open: false.", "pointerdown on select at wrapper. open: false.", "mousedown on select at select. open: false.", "mousedown on select at wrapper. open: false.", "focusin on select at select. open: false.", "focusin on select at wrapper. open: false.", "pointerup on select at select. open: true.", "pointerup on select at wrapper. open: true.", "mouseup on select at select. open: true.", "mouseup on select at wrapper. open: true.", "click on select at select. open: true.", "click on select at wrapper. open: true."])
FAIL Events, explicit button promise_test: Unhandled rejection with value: object "Error: element click intercepted error"
